model/requirements.txt
flask==3.1.0
pandas==2.2.3
mlxtend==0.23.1
python-dotenv==1.0.1